requirements.txt
boto3>=1.34.136,<2
GitPython>=3.1.43,<4
requests>=2.32.3,<3
watcloud-utils @ git+https://github.com/WATonomous/watcloud-utils.git@c8ce1006716e65971f750560f90f442721b3777d
jsonschema>=4.23.0,<5

# For testing
moto[s3]>=5,<6
pytest>=8,<9